refactor: rename schema to schemas
