+++ api/package.json
@@ -4,10 +4,10 @@
   "main": "index.js",
   "private": true,
   "scripts": {
-    "watch:server": "nodemon --exec 'ts-node -r tsconfig-paths/register src/index.ts'",
-    "watch:codegen": "nodemon --exec 'yarn codegen' --watch graphql/** --ext graphql",
+    "watch:server": "nodemon --exec 'ts-node -r tsconfig-paths/register src/index.ts' --ignore __generated --ext .ts",
+    "watch:codegen": "nodemon --exec 'yarn codegen' --watch schemas/** --ext graphql",
     "lint": "eslint --ext .ts .",
-    "codegen:schema": "ts-node -r tsconfig-paths/register src/build-schema.ts",
+    "codegen:schema": "ts-node -r tsconfig-paths/register --transpile-only src/build-schema.ts",
     "codegen:types": "graphql-codegen --config codegen.ts",
     "codegen": "yarn codegen:schema && yarn codegen:types",
     "postinstall": "yarn codegen"
